Added make rules ~ todo: fix loading lib
