+++ .gitignore
@@ -5,3 +5,4 @@
 *.o
 
 lib/
+main
