+++ src/dynamic_linkage.hpp
@@ -1,6 +1,10 @@
-#ifndef DYNAMIC_HPP
-#define DYNAMIC_HPP
+#ifndef DYNAMIC_LINKAGE_HPP
+#define DYNAMIC_LINKAGE_HPP
 
-void print_random_name();
+#include <iostream>
 
-#endif // DYNAMIC_HPP
+unsigned int power(unsigned int x, unsigned int p);
+
+void print_inversion(unsigned int x);
+
+#endif // DYNAMIC_LINKAGE_HPP
