+++ src/dynamic_load.hpp
@@ -1,10 +1,6 @@
-#ifndef DYNAMIC_LINKAGE_HPP
-#define DYNAMIC_LINKAGE_HPP
+#ifndef DYNAMIC_HPP
+#define DYNAMIC_HPP
 
-#include <iostream>
+void print_random_name();
 
-unsigned int power(unsigned int x, unsigned int p);
-
-void print_inversion(unsigned int x);
-
-#endif // DYNAMIC_LINKAGE_HPP
+#endif // DYNAMIC_HPP
